Python/pyproject.toml
[build-system]
requires = ["scikit-build-core", "pybind11"]
build-backend = "scikit_build_core.build"


[project]
name = "taospyudf"
version = "0.0.3"
description="taos python udf"
readme = "README.md"
authors = [
  { name = "taos data", email = "slzhou@taosdata.com" },
]
requires-python = ">=3.7"
classifiers = [
  "Development Status :: 4 - Beta",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
]

[project.optional-dependencies]
test = ["pytest", "taospy"]


[tool.scikit-build]
wheel.expand-macos-universal-tags = true
sdist.exclude = ["extern/pybind11", "tests/__pycache__", "tests/.pytest_cache"]


[tool.cibuildwheel]
test-command = "pytest {project}/tests"
test-extras = ["test"]
test-skip = ["*universal2:arm64"]
build-verbosity = 1
